requirements.txt
numpy>=2.0.0,<2.1.0
pandas==2.2.2
torch==2.5.1
psutil==5.9.0
transformers==4.37.2
pydub==0.25.1
python-dotenv==1.0.0
gtts==2.5.3
moviepy==1.0.3
Pillow==10.0.1
spacy==3.8.2
openai-whisper==20240930
